feat: 문자열 배열을 일급 컬렉션 Source로 저장하는 기능 구현
