+++ README.md
@@ -5,7 +5,7 @@
 - [x] 문자열 입력을 받는 기능
 - [x] 공백을 기준으로 문자열을 split하는 기능
 - [x] 문자열 배열의 원소 개수가 홀수인지 검증하는 기능
-- [ ] 문자열 배열을 일급 컬렉션 Source로 저장
+- [x] 문자열 배열을 일급 컬렉션 Source로 저장
     - 인덱스가 홀수일 때, 숫자인지 검증하고 operands 리스트에 추가
     - 인덱스가 짝수일 때, 사칙연산기호인지 검증하고 operators 리스트에 추가
 - [ ] 위의 세가지를 검증 후 예외인 경우 처리하는 기능
